search modification
Use of g_timeout_add to delay the search at each inputbar modification
so we avoid search on the first characters of the pattern.
We don't re-highlight the pattern at each launch of cmd_search.

Former behavior are kept except : if I search something and then hit Esc
the highlight disappear, then if I launch cmd_search the pattern will be
search but the highlight will not reappear.

Signed-off-by: Maxime de Roucy <maxime.deroucy@gmail.com>